retrofit(flask_ver)연동 완료_근데 이제 파싱문제가 좀 있는;
