+++ app/src/main/res/values/strings.xml
@@ -3,6 +3,5 @@
     <string name="title_map">Map</string>
     <string name="title_home">Home</string>
     <string name="title_makecourse">Course maker</string>
-    <string name="baseUrl">http://127.0.0.1:5000</string>
-
+    <string name="baseUrl">http://192.168.35.180:5000</string>
 </resources>
